Scoped display

+++ crates/assets/src/assets/engine/shaders/gui.vrsh.glsl
@@ -1,11 +1,13 @@
 // https://github.com/cohaereo/egui_glfw_gl/blob/master/src/painter.rs
 #version 460 core
-uniform ivec2 resolution;
+uniform vec2 resolution;
 layout(location = 0) in vec2 a_pos;
 layout(location = 1) in vec2 a_tc;
 layout(location = 2) in vec4 a_srgba; // 0-255 sRGB
 out vec4 v_rgba;
 out vec2 v_tc;
+
+// TODO: Is there a way to remove all of this for a simpler solution?
 // 0-1 linear  from  0-255 sRGB
 vec3 linear_from_srgb(vec3 srgb) {
     bvec3 cutoff = lessThan(srgb, vec3(10.31475));
